authorization

+++ app/Http/Controllers/IdeaController.php
@@ -21,6 +21,8 @@ class IdeaController extends Controller
         $validated = request() -> validate([
             'content' => 'required|min:5|max:240'
         ]);
+
+        $validated['user_id'] = auth() -> id();
         
         Idea::create($validated);
 
@@ -30,6 +32,10 @@ class IdeaController extends Controller
 
     public function destroy(Idea $id){
 
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
        $idea->delete();
 
        
@@ -38,12 +44,20 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea){
 
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         $editing = true;
 
         return view('ideas.show',compact('idea','editing'));
     }
 
     public function update(Idea $idea){
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
 
         request() -> validate([
             'content' => 'required|min:5|max:240'
